城用のアイコン追加、城のHP減少処理修正、プレイヤーのシーン用演出完成。

+++ Tank_Ojousama/DirectX12/Actor/Castle.h
@@ -35,7 +35,7 @@ private:
 	//–³“GŽžŠÔ‚Ì‚½‚ß
 	bool HitFlag;
 	int HitCount;
-	float MaxHp;
+	int MaxHp;
 	int MessCount1;
 	int MessCount2;
 	int MessCount3;
